moved "languages" from browser_languages package to pages package and moved "languages" to package tests
