+++ src/test/java/pages/languages/A_Page.java
@@ -1,4 +1,4 @@
-package pages.browse_languages.languages;
+package pages.languages;
 
 
 import org.openqa.selenium.By;
