+++ src/test/java/pages/languages/CSharp1614Page.java
@@ -1,4 +1,4 @@
-package pages.browse_languages.languages;
+package pages.languages;
 
 import org.openqa.selenium.WebDriver;
