+++ src/test/java/pages/top_lists/TopListPage.java
@@ -2,7 +2,7 @@ package pages.top_lists;
 
 import org.openqa.selenium.WebDriver;
 
-import pages.browse_languages.languages.LanguagesPage;
+import pages.languages.LanguagesPage;
 
 public class TopListPage extends BaseTopListPage<LanguagesPage> {
