+++ src/test/java/tests/languages/Algol601367Test.java
@@ -1,8 +1,8 @@
-package tests.browse_languages.languages;
+package tests.languages;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;
-import pages.browse_languages.languages.LanguagesPage;
+import pages.languages.LanguagesPage;
 import runner.BaseTest;
 
 public class Algol601367Test extends BaseTest {
