+++ src/test/java/tests/languages/AlphaBetaTest.java
@@ -1,8 +1,8 @@
-package tests.browse_languages.languages;
+package tests.languages;
 
 import org.testng.Assert;
 import org.testng.annotations.Test;
-import pages.browse_languages.languages.AlphaBetaPage;
+import pages.languages.AlphaBetaPage;
 import runner.BaseTest;
 
 public class AlphaBetaTest extends BaseTest {
