+++ src/test/java/tests/languages/BeginWithZTest.java
@@ -1,4 +1,4 @@
-package tests.browse_languages.languages;
+package tests.languages;
 
 import org.testng.annotations.Test;
 import runner.BaseTest;
